Have got an unofficial Log Page, awaiting group-collaboration to finish it off
I've got a headache, but the short version is that Logs can now be sorted out via userID and date, and I've written up a viewing page for the logs. Just need a controller for the connection between the model and view and a way to seperate admins from regulars and to create logs on session end. Lovely.