Now have a login and logout independant way of creating logs (I think)
Both the logpage and createlog link to each other, so hopefully we can create logs as an admin. It wouldn't be a permanent feature, being merely there so that we can demonstrate the admin side of logs working.

In doing this, I also remembered that this is technically Release 1, and that the full project would feature more complete code.